backend/infrastructure/requirements.txt
# CDK Infrastructure Dependencies
aws-cdk-lib>=2.213.0
constructs>=10.4.2
